#185-refactor internal states implementations
Fully completed. Validated.
