+++ c++20/baseclasses/baselfib64.h
@@ -31,7 +31,7 @@ SOFTWARE.
 #include <cstdint>
 
 #include "baserandom.h"
-#include "listseedstate.h"
+#include "internalstates/listseedstate.h"
 #include "utils/splitmix.h"
