+++ c++20/baseclasses/basexoroshiro.h
@@ -31,7 +31,7 @@ SOFTWARE.
 #include <array>
 
 #include "baserandom.h"
-#include "listseedstate.h"
+#include "internalstates/listseedstate.h"
 #include "utils/splitmix.h"
