+++ c++20/mrg1457.h
@@ -30,7 +30,7 @@ SOFTWARE.
 #include <type_traits>
 
 #include "baseclasses/basemrg31.h"
-#include "listseedstate.h"
+#include "internalstates/listseedstate.h"
 
 
 //===========================================================================
