+++ c++20/mrg287.h
@@ -31,7 +31,7 @@ SOFTWARE.
 #include <type_traits>
 
 #include "baseclasses/basemrg32.h"
-#include "listseedstate.h"
+#include "internalstates/listseedstate.h"
 
 
 //===========================================================================
