+++ c++20/mrg49507.h
@@ -31,7 +31,7 @@ SOFTWARE.
 #include <type_traits>
 
 #include "baseclasses/basemrg31.h"
-#include "listseedstate.h"
+#include "internalstates/listseedstate.h"
 
 
 //===========================================================================
